Pluginify summary aggregation stage (#3817)
